Adds test_switch_to_repeated_name_screen() and the ScreenManagerTestCase class in the renamed file test_screen_manager.py
